# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  coverImage: string;
  publishedAt: string;
  readTimeMinutes: number;
  tags: string[];
  content: string;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  excerpt: string;
  coverImage: string;
  publishedAt: string;
  /** @minimum 1 */
  readTimeMinutes: number;
  tags: string[];
  content: string;
  isPublished?: boolean;
}

export interface UpdatePostInput {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  title?: string;
  excerpt?: string;
  coverImage?: string;
  publishedAt?: string;
  /** @minimum 1 */
  readTimeMinutes?: number;
  tags?: string[];
  content?: string;
  isPublished?: boolean;
}

export type ProjectCategory = typeof ProjectCategory[keyof typeof ProjectCategory];


export const ProjectCategory = {
  pcb: 'pcb',
  iot: 'iot',
  esp32: 'esp32',
  arm: 'arm',
  arduino: 'arduino',
  simulation: 'simulation',
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  details: string;
  category: ProjectCategory;
  image: string;
  tags: string[];
  link: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
  details?: string;
  image: string;
  tags: string[];
  link: string;
  category: ProjectCategory;
  sortOrder?: number;
}

export interface UpdateProjectInput {
  /** @minLength 1 */
  title?: string;
  description?: string;
  details?: string;
  category?: ProjectCategory;
  image?: string;
  tags?: string[];
  link?: string;
  sortOrder?: number;
}

export interface Section {
  key: string;
  value: string;
  updatedAt: string;
}

export interface UpsertSectionInput {
  value: string;
}

export interface Skill {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  level: number;
}

export interface SkillCategory {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  title: string;
  /**
     * Lucide icon name used by the frontend (e.g. Cpu, CircuitBoard).
     * @minLength 1
     */
  icon: string;
  skills: Skill[];
}

export interface SkillsConfig {
  categories: SkillCategory[];
}

export interface SendContactMessageInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 5 */
  subject: string;
  /** @minLength 10 */
  message: string;
  /** Honeypot field (should be empty). */
  company?: string;
}

export interface ContactMessageResult {
  ok: boolean;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface UpdateContactMessageInput {
  isRead: boolean;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListPostsParams = {
includeUnpublished?: boolean;
};

export type ListContactMessagesParams = {
unreadOnly?: boolean;
};