# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
...
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
...
  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List blog posts
      parameters:
        - name: includeUnpublished
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Array of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a new blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /posts/by-slug/{slug}:
    get:
      tags: [Posts]
      operationId: getPostBySlug
      summary: Get a post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /posts/{id}:
    get:
      tags: [Posts]
      operationId: getPost
      summary: Get a post by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Posts]
      operationId: updatePost
      summary: Update a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List portfolio projects
      responses:
        '200':
          description: Array of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a new project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /projects/{id}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a project by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /contact-messages:
    get:
      tags: [Contact]
      operationId: listContactMessages
      summary: List contact form messages (admin)
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Array of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /contact-messages/{id}:
    get:
      tags: [Contact]
      operationId: getContactMessage
      summary: Get a contact message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Contact]
      operationId: updateContactMessage
      summary: Update a contact message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactMessageInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Contact]
      operationId: deleteContactMessage
      summary: Delete a contact message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Post:
      type: object
      required: [id, slug, title, excerpt, coverImage, publishedAt, readTimeMinutes, tags, content, isPublished, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        coverImage:
          type: string
        publishedAt:
          type: string
        readTimeMinutes:
          type: integer
        tags:
          type: array
          items:
            type: string
        content:
          type: string
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePostInput:
      type: object
      required: [slug, title, excerpt, coverImage, publishedAt, readTimeMinutes, tags, content]
      properties:
        slug:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        excerpt:
          type: string
        coverImage:
          type: string
        publishedAt:
          type: string
        readTimeMinutes:
          type: integer
          minimum: 1
        tags:
          type: array
          items:
            type: string
        content:
          type: string
        isPublished:
          type: boolean
    UpdatePostInput:
      type: object
      properties:
        slug:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        excerpt:
          type: string
        coverImage:
          type: string
        publishedAt:
          type: string
        readTimeMinutes:
          type: integer
          minimum: 1
        tags:
          type: array
          items:
            type: string
        content:
          type: string
        isPublished:
          type: boolean
    Project:
      type: object
      required: [id, title, description, details, image, tags, link, sortOrder, createdAt, updatedAt, category]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        details:
          type: string
        category:
          $ref: '#/components/schemas/ProjectCategory'
        image:
          type: string
        tags:
          type: array
          items:
            type: string
        link:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectInput:
      type: object
      required: [title, description, image, tags, link, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        details:
          type: string
        image:
          type: string
        tags:
          type: array
          items:
            type: string
        link:
          type: string
        category:
          $ref: '#/components/schemas/ProjectCategory'
        sortOrder:
          type: integer
    UpdateProjectInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        details:
          type: string
        category:
          $ref: '#/components/schemas/ProjectCategory'
        image:
          type: string
        tags:
          type: array
          items:
            type: string
        link:
          type: string
        sortOrder:
          type: integer
...
    ProjectCategory:
      type: string
      enum:
        - pcb
        - iot
        - esp32
        - arm
        - arduino
        - simulation
...
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, isRead, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        subject:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateContactMessageInput:
      type: object
      required: [isRead]
      properties:
        isRead:
          type: boolean