# lib/api-zod/src/generated/types/createPostInput.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  excerpt: string;
  coverImage: string;
  publishedAt: string;
  /** @minimum 1 */
  readTimeMinutes: number;
  tags: string[];
  content: string;
  isPublished?: boolean;
}