# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectCategory } from './projectCategory';

export interface CreateProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
  details?: string;
  image: string;
  tags: string[];
  link: string;
  category: ProjectCategory;
  sortOrder?: number;
}